# ⚡️ Speed up function `group_broken_paragraphs` by 30%