--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,15 @@
 
 ### Fixes
 
+## 0.18.15-dev0
+
+### Enhancements
+- Optimized the runtime of `ElementHtml._get_children_html`
+
+### Features
+
+### Fixes
+
 ## 0.18.14
 
 ### Enhancements