@aseembits93 2025-09-05
@qued everything passing except the docker CI, logs seem to indicate `poppler-utils` not present in the system. I am facing this issue in another PR of mine (https://github.com/Unstructured-IO/unstructured/pull/4087#issuecomment-3250821307). Hope to get this resolved! let me know if there's anything I could do from my end, Best,